refactor: extract ExportResult.swift from IssueExportTargets.swift
Splits the `ExportResult` value (post-export receipt data) out of
IssueExportTargets.swift so that file focuses on target payloads.

Byte-identical move of lines 100-124. Both files import Foundation.

IssueExportTargets.swift: 124 → 98 lines. Tests: 667.
